% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_test.R
\name{boot_test}
\alias{boot_test}
\title{Bootstrap Confidence Intervals via Resampling}
\usage{
boot_test(x, y, conf.int, resample, med)
}
\arguments{
\item{x, y}{numeric vectors of data values}

\item{conf.int}{(optional) confidence level of the interval. Defaults to \code{0.90}}

\item{resample}{(optional) number of resamples. Defaults to 10,000}

\item{med}{(optional) number indicating true difference in medians to test against. Defaults to zero.}
}
\description{
Provides nonparametric confidence intervals via percentile-based resampling.
}
\details{
Refer to vignette for further information.
}
\examples{
require(graphics)

a <- rnorm(25, 80, 35)
b <- rnorm(25, 100, 50)

boot_test(a, b, 0.95, 10000)

}
