% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.R
\name{corr}
\alias{corr}
\title{Correlation Coefficient}
\usage{
corr(r, n, conf.int)
}
\arguments{
\item{r}{correlation coefficient}

\item{n}{sample size}

\item{conf.int}{(optional) confidence level of the interval. Defaults to \code{0.90}}
}
\description{
Provides magnitude-based inferences upon given \emph{r} value and sample size. Based upon WG Hopkins Microsoft Excel spreadsheet.
}
\details{
Refer to vignette for further information.
}
\examples{
corr(.40, 25, 0.95)
}
\references{
Hopkins WG. (2007). A spreadsheet for deriving a confidence interval, mechanistic inference and clinical inference from a \emph{p} value. \emph{Sportscience} 11, 16-20. sportsci.org/2007/wghinf.htm
}
