% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_diff.R
\name{corr_diff}
\alias{corr_diff}
\title{Test of Two Correlations}
\usage{
corr_diff(r1, n1, r2, n2, conf.int = 0.9, plot = FALSE)
}
\arguments{
\item{r1}{correlation of group 1}

\item{n1}{sample size of group 1}

\item{r2}{correlation of group 2}

\item{n2}{sample size of group 2}

\item{conf.int}{(optional) confidence level of the interval. Defaults to \code{0.90}}

\item{plot}{(optional) logical indicator specifying to print associated plot. Defaults to \code{FALSE}}
}
\description{
Provides statistical inference upon the difference between two independent correlations.
}
\details{
Refer to vignette for further information.
}
\examples{
corr_diff(r1 = 0.20, n1 = 71, r2 = 0.55, n2 = 46)
}
\references{
Zou GY. (2007). Toward using confidence intervals to compare correlations. \emph{Psychological Methods}, 12, 399-413.
}
