% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datambicorr.h.R
\name{dataMBIcorr}
\alias{dataMBIcorr}
\title{MBI for Correlation}
\usage{
dataMBIcorr(data, pairs, pearson = TRUE, spear = FALSE, kendall = FALSE,
  SWC = 0.1, confint = 90, desc = FALSE)
}
\arguments{
\item{data}{the data as a data frame}

\item{pairs}{a list of vectors of strings naming variables to pair from
\code{data}}

\item{pearson}{\code{TRUE} (default) or \code{FALSE}, provide Pearson's R}

\item{spear}{\code{TRUE} or \code{FALSE} (default), provide Spearman's rho}

\item{kendall}{\code{TRUE} or \code{FALSE} (default), provide Kendall's
tau-b}

\item{SWC}{a number, default (r=0.1) that defines the bounds for
magnitude-based inference}

\item{confint}{90\% (default), confidence interval for magnitude-based
inference}

\item{desc}{\code{TRUE} or \code{FALSE} (default), provide descriptive
statistics}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$mbi} \tab \tab \tab \tab \tab a table \cr
  \code{results$desc} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$mbi$asDF}

\code{as.data.frame(results$mbi)}
}
\description{
MBI for Correlation
}
\examples{
\dontrun{

data('bugs', package = 'jmv')

mbir::dataMBIcorr(
    data = data,
    pairs = list(
        list(
            i1="LDLF",
            i2="LDHF")))

#
#  Magnitude Based Inference: Correlation
#  -------------------------------------------------------------------------------------------------------------------------------------------------
#                                            p         Negative (\\\%)    Trivial (\\\%)    Positive (\\\%)    Inference
#  -------------------------------------------------------------------------------------------------------------------------------------------------
#    LDLF    LDHF    Pearson's r    0.577    < .001            0.00           0.00             100    Almost Certainly Large Positive Correlation.
#  -------------------------------------------------------------------------------------------------------------------------------------------------
#
#
}
}
