% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datambinpind.h.R
\name{dataMBInpind}
\alias{dataMBInpind}
\title{MBI for Independent Samples (non-parametric)}
\usage{
dataMBInpind(data, deps, group, SWC = 0.5, confint = 90, desc = FALSE)
}
\arguments{
\item{data}{the data as a data frame}

\item{deps}{a vector of strings naming the dependent variables}

\item{group}{a string naming the grouping variable, must have 2 levels}

\item{SWC}{a number that defines the bounds for magnitude-based inference}

\item{confint}{90\% (default) confidence intervals for effect sizes}

\item{desc}{\code{TRUE} or \code{FALSE} (default), provide descriptive
statistics}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$ttest} \tab \tab \tab \tab \tab a table \cr
  \code{results$effect} \tab \tab \tab \tab \tab a table \cr
  \code{results$mbi} \tab \tab \tab \tab \tab a table \cr
  \code{results$desc} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$ttest$asDF}

\code{as.data.frame(results$ttest)}
}
\description{
MBI for Independent Samples (non-parametric)
}
\examples{
\dontrun{

data('bugs', package = 'jmv')

mbir::dataMBInpind(
data = bugs,
deps = "LDLF",
group = "Gender")
}
}
