% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbi_corr.R
\name{mbi_corr}
\alias{mbi_corr}
\title{Magnitude Based Inference for a Correlation}
\usage{
mbi_corr(
  x,
  y,
  conf.level = mbir_options("conf.level"),
  method = "pearson",
  low_eqbound = -0.1,
  high_eqbound = 0.1,
  mech_decisions = mbir_options("mech_decisions"),
  plot = mbir_options("plot")
)
}
\arguments{
\item{x, y}{numeric vectors of data values}

\item{method}{(character) if \code{auto = F}, logical indicator specifying which correlation to execute (\code{pearson, spearman, kendall}). Defaults to \code{"pearson"}.}

\item{plot}{(optional) logical indicator specifying to print associated plot. Defaults to \code{FALSE}}

\item{conf.int}{(optional) confidence level of the interval. Defaults to \code{0.90}}

\item{auto}{(character) logical indicator specifying if user wants function to programmatically detect statistical procedures. Defaults to \code{TRUE}}

\item{swc}{(optional) number indicating smallest worthwhile change. Defaults to \code{0.1}}
}
\value{
Associated effect size measure, \emph{r}, and respective confidence intervals.
}
\description{
Provides magnitude-based inferences for the association between given data vectors. Evaluates normality assumption, performs either Pearson or Spearman correlation and subsequently estimates magnitude-based inferences.
}
\details{
Refer to vignette for further information.
}
\examples{
a <- rnorm(25, 80, 35)
b <- rnorm(25, 100, 35)

mbi_corr(a, b)
}
