% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbi_corr_rev.R
\name{mbi_corr_rev}
\alias{mbi_corr_rev}
\title{Correlation Coefficient}
\usage{
mbi_corr_rev(
  r,
  n,
  conf.level = 0.95,
  low_eqbound = -0.1,
  high_eqbound = 0.1,
  mech_decisions = list(strong_alpha = 0.005, moderate_alpha = 0.05, weak_alpha = 0.25),
  plot = mbir_options("plot"),
  verbose = mbir_options("verbose")
)
}
\arguments{
\item{r}{Pearson correlation coefficient}

\item{n}{sample size (number of pairs)}

\item{conf.level}{Confidence interval level. Defaults to \code{0.95}}

\item{low_eqbound}{Lower equivalence bounds (e.g., -0.1) expressed in raw units}

\item{high_eqbound}{Upper equivalence bounds (e.g., 0.1) expressed in raw units}

\item{mech_decisions}{List containing options for mechanistic inference. Options include strong_alpha, mod_alpha, and weak_alpha.}

\item{plot}{(optional) logical indicator specifying to print associated plot. Defaults to \code{FALSE}}
}
\description{
Provides magnitude-based inferences upon given \emph{r} value and sample size. Unlike the old functions, this follows the hypothesis testing guidelines set forth by
}
\details{
This function back calculates multiple one-tailed t-tests in order to peform equivalence and minimal effects tests for a correlation. All testing and confidence intervals are calculated through Fisher's z-transformation. Results will differ from the stats cor.test function which utilizes a t-transformation.
}
\examples{
mbi_corr_rev(.40, 25, 0.95)
}
