% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbi_mu_data.R
\name{mbi_mu}
\alias{mbi_mu}
\title{MBI for Independent or Dependent Groups Comparison}
\usage{
mbi_mu(
  x,
  y,
  paired = TRUE,
  var.equal = TRUE,
  low_eqbound,
  high_eqbound,
  inference = "mechanistic",
  conf.level = mbir_options("conf.level"),
  mech_decisions = mbir_options("mech_decisions"),
  clin_decisions = mbir_options("clin_decisions"),
  plot = mbir_options("plot"),
  verbose = mbir_options("verbose")
)
}
\arguments{
\item{x, y}{numeric vectors of data values}

\item{paired}{logical indicator specifying if \code{x} and \code{y} are paired \code{(TRUE)} or independent \code{(FALSE)}}

\item{var.equal}{(optional) if \code{auto = F}, logical indicator specifying if homogeneity of variance assumed. Defaults to \code{TRUE}}

\item{low_eqbound}{Lower equivalence bounds (e.g., -0.5) expressed in raw units}

\item{high_eqbound}{Upper equivalence bounds (e.g., 0.5) expressed in raw units}

\item{inference}{The type of MBI hypothesis testing to be performed; options are clinical or mechanistic.}

\item{conf.level}{Confidence level of the interval estimate. Default is 95\%, Ex: conf.level = 0.}

\item{mech_decisions}{List containing options for mechanistic inference. Options include strong_alpha, mod_alpha, and weak_alpha.}

\item{clin_decisions}{List containing options for clinical inference. Options include the use_direction (the direction of the effect that is considered positive or good), harm_alpha and benefit_alpha.}

\item{plot}{(optional) logical indicator specifying to print associated plot. Defaults to \code{FALSE}}

\item{verbose}{logical indicator that when \code{(TRUE)} prints inference to the console.}
}
\value{
Associated effect size measures (\emph{d}, \emph{r}, odds ratio) and respective confidence intervals based upon which statistical test(s) performed.
}
\description{
Performs two-sample difference of means analysis to produce magnitude-based inferences hypothesis tests.
}
\details{
This function calculates multiple one-tailed t-tests in order to peform equivalence and minimal effects tests and interprets the data based on magnitude based inference hypothesis testing scheme.
}
\references{
Aisbett, J., Lakens, D., & Sainani, K. (2020, May 3). Magnitude Based Inference in Relation to One-sided Hypotheses Testing Procedures. https://doi.org/10.31236/osf.io/pn9s3
}
