% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbi_mu_rev.R
\name{mbi_mu_rev}
\alias{mbi_mu_rev}
\title{MBI for Mean Difference Reverse Engineered from Summary Statistics}
\usage{
mbi_mu_rev(
  point,
  LL = NULL,
  UL = NULL,
  se = NULL,
  n,
  low_eqbound,
  high_eqbound,
  paired = TRUE,
  conf.level = mbir_options("conf.level"),
  mech_decisions = mbir_options("mech_decisions"),
  clin_decisions = mbir_options("clin_decisions"),
  inference = "mechanistic",
  plot = mbir_options("plot"),
  verbose = mbir_options("verbose")
)
}
\arguments{
\item{point}{The point estimate of the mean difference from the analysis. Ex: 1.20}

\item{LL}{The lower confidence limit from an analysis. Optional: either se or confidence limits (UL & LL) must be entered Ex: LL = 1.0}

\item{UL}{The upper confidence limit from an analysis. Optional: either se or confidence limits (UL & LL) must be entered Ex: UL = 1.4}

\item{se}{The standard error of the point estimate. Optional: either se or confidence limits (UL & LL) must be entered. Ex: se = 0.05}

\item{n}{Sample size; vector of sample sizes for independent samples or number of pairs for paired samples.}

\item{low_eqbound}{Lower equivalence bounds (e.g., -0.5) expressed in raw units}

\item{high_eqbound}{Upper equivalence bounds (e.g., 0.5) expressed in raw units}

\item{paired}{logical indicator specifying if comparison is from paired \code{(TRUE)} or independent \code{(FALSE)} samples.}

\item{conf.level}{Confidence level of the interval estimate. Default is 95\%, Ex: conf.level = 0.95}

\item{mech_decisions}{List containing options for mechanistic inference. Options include strong_alpha, mod_alpha, and weak_alpha.}

\item{clin_decisions}{List containing options for clinical inference. Options include the use_direction (the direction of the effect that is considered positive or good), harm_alpha and benefit_alpha.}

\item{inference}{The type of MBI hypothesis testing to be performed; options are clinical or mechanistic}

\item{plot}{logical indicator that when \code{(TRUE)} returns a consonance plot.}

\item{verbose}{logical indicator that when \code{(TRUE)} prints inference to the console.}
}
\value{
Returns a consonance plot (cons_plot), results from the MBI associated t-tests (result_table), and the resulting MBI hypothesis testing inference (inference)
}
\description{
Provides magnitude-based interpretation using summary statistic (mean and standard error or confidence limits). Plots are based upon the functions in the \code{(concurve)} R package.
}
\details{
This function back calculates multiple one-tailed t-tests in order to peform equivalence and minimal effects tests.
}
\references{
Aisbett, J., Lakens, D., & Sainani, K. (2020, May 3). Magnitude Based Inference in Relation to One-sided Hypotheses Testing Procedures. https://doi.org/10.31236/osf.io/pn9s3
}
