% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbir_options.R
\name{mbir_options}
\alias{mbir_options}
\title{Set/get global mbir options}
\usage{
mbir_options(...)
}
\arguments{
\item{...}{One of four: (1) nothing, then returns all options as a list; (2)
a name of an option element, then returns its' value; (3) a name-value pair
which sets the corresponding option to the new value (and returns nothing),
(4) a list with option-value pairs which sets all the corresponding
arguments. The example show all possible cases.}
}
\value{
depends on input, see above.
}
\description{
Global mbir options are used by many functions. But can be changed in each functions directly using
an argument (which has precedence over the global options).
}
\details{
The following arguments are currently set:
\itemize{
\item \code{verbose} should verbose (printed results) be set to true? Default is \code{TRUE}.
\item \code{plot} Option to automatically print plots. Default is \code{FALSE}.
}
}
\note{
All options are saved in the global R \code{\link{options}} with prefix
  \code{mbir.}
}
