% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds.R
\name{odds}
\alias{odds}
\title{Odds Ratio}
\usage{
odds(or, p, conf.int = 0.9)
}
\arguments{
\item{or}{odds ratio}

\item{p}{associated \emph{p}-value}

\item{conf.int}{(optional) confidence level of the interval. Defaults to \code{0.90}}
}
\description{
Provides magnitude-based inferences upon given odds ratio and \emph{p}-value. Based upon WG Hopkins Microsoft Excel spreadsheet.
}
\details{
Refer to vignette for further information.
}
\examples{
odds(1.25, 0.06, 0.95)
}
\references{
Hopkins WG. (2007). A spreadsheet for deriving a confidence interval, mechanistic inference and clinical inference from a \emph{p} value. \emph{Sportscience} 11, 16-20. sportsci.org/2007/wghinf.htm
}
