% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop.R
\name{prop}
\alias{prop}
\title{Test of Two Proportions}
\usage{
prop(p1, n1, p2, n2, conf.int)
}
\arguments{
\item{p1}{proportion of group 1}

\item{n1}{sample size of group 1}

\item{p2}{proportion of group 2}

\item{n2}{sample size of group 2}

\item{conf.int}{(optional) confidence level of the interval. Defaults to \code{0.90}}
}
\description{
Provides magnitude-based inferences upon given proportions and sample sizes. Based upon WG Hopkins Microsoft Excel spreadsheet.
}
\details{
Refer to vignette for further information.
}
\examples{
prop(p1 = 0.7, n1 = 25, p2 = 0.5, n2 = 20)
}
\references{
Hopkins WG. (2007). A spreadsheet for deriving a confidence interval, mechanistic inference and clinical inference from a \emph{p} value. \emph{Sportscience} 11, 16-20. sportsci.org/2007/wghinf.htm
}
