% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_corr.R
\name{ss_corr}
\alias{ss_corr}
\title{Sample Size Estimation: Correlation Coefficient}
\usage{
ss_corr(n, r)
}
\arguments{
\item{n}{planned sample size}

\item{r}{planned correlation coefficient}
}
\description{
Estimates magnitude-based inferences upon planned sample size and \emph{r} value. Based upon WG Hopkins Microsoft Excel spreadsheet.
}
\details{
Refer to vignette for further information.
}
\examples{
ss_corr(n = 20, r = 0.2)
}
\references{
Hopkins WG. (2006). Estimating sample size for magnitude-based inferences. \emph{Sportscience} 10, 63-70. sportsci.org/2006/wghss.htm
}
