% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_odds.R
\name{ss_odds}
\alias{ss_odds}
\title{Sample Size Estimation: Odds Ratio}
\usage{
ss_odds(exp, con, or)
}
\arguments{
\item{exp}{planned sample size of experimental group}

\item{con}{planned sample size of control group}

\item{or}{planned odds ratio}
}
\description{
Estimates magnitude-based inferences upon planned sample size and odds ratio. Based upon WG Hopkins Microsoft Excel spreadsheet.
}
\details{
Refer to vignette for further information.
}
\examples{
ss_odds(exp = 15, con = 18, or = 3.25)
}
\references{
Hopkins WG. (2006). Estimating sample size for magnitude-based inferences. \emph{Sportscience} 10, 63-70. sportsci.org/2006/wghss.htm
}
