% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_smd.R
\name{ss_smd}
\alias{ss_smd}
\title{Sample Size Estimation: Standardized Mean Difference}
\usage{
ss_smd(exp, con, es)
}
\arguments{
\item{exp}{planned sample size of experimental group}

\item{con}{planned sample size of control group}

\item{es}{planned Cohen's \emph{d}}
}
\description{
Estimates magnitude-based inferences upon planned sample size and \emph{d} value. Based upon WG Hopkins Microsoft Excel spreadsheet.
}
\details{
Refer to vignette for further information.
}
\examples{
ss_smd(exp = 20, con = 15, es = 0.6)
}
\references{
Hopkins WG. (2006). Estimating sample size for magnitude-based inferences. \emph{Sportscience} 10, 63-70. sportsci.org/2006/wghss.htm
}
